Small typos and todo's

+++ simulator/src/config.h
@@ -21,7 +21,7 @@ public:
     static const double DT;// simulator will update at 0.5 seconds.
 
     // how far from the end of the road the traffic light should be positioned
-    static const unsigned trafficLightDistToRoadEnd; // = 3 meters
+    static const double trafficLightDistToRoadEnd; // = 1 meters
 };
 
 } // namespace simulator
